God save the ... King?
The British are going to get used to a whole host of changes after Queen Elizabeth II's death. Pretty much most of them have known life under only one monarch. But with the ascension of King Charles III, even the country's national anthem will now have to change.

It's worth noting each country within the United Kingdom has its own anthem — Scotland uses "Flower of Scotland," for example, at sporting events — but England's "God Save the Queen" has been taken up as the signature song for Britain.

Originally titled "God Save the King," the song is based on a patriotic tune that was first publicly performed in London in 1745 and was adopted as the country's national anthem under the reign of George II in the 18th century. The words and melody have no certain origin, and typically only the first verse is sung.

For now, any official instructions are pending given that the royal.uk website has gone dark. On the subject of national anthems, the website says: "The words used today are those sung in 1745, substituting 'Queen' for 'King' where appropriate."

Now that King Charles III's reign has begun, the anthem's changes should be in effect at any event where the song is usually performed, a spokesperson for Debrett's, a leading guide on British etiquette, told The Times in an email Thursday.

The song was also adapted to "God Save the Queen" for Queen Victoria, according to Debrett's.

The "God Save the King" version was then sung until 1952 when Elizabeth II's father, George VI, died and she became queen.

As the U.K. enters a period of national mourning, with many events being canceled ahead of the queen's funeral, it is not clear when we will hear the first public chorus of the anthem.

After her death was announced Thursday, some standing vigil outside Buckingham Palace sang the new version, according to a CNN reporter.

At one soccer match Thursday evening in England, fans burst into the old version of the anthem when the queen's death was announced to the crowd.

The first major moment for the new national anthem would have been on Saturday at the Last Night of the Proms — the British classical music festival that culminates in a final night with a rousing rendition of the anthem. As a mark of respect to the queen's passing, that event was canceled.

For when the time comes to sing the anthem again, here are the words:

God save our gracious King!

Long live our noble King!

God save the King!

Send him victorious,

Happy and glorious,

Long to reign over us,

God save the King

Thy choicest gifts in store

On him be pleased to pour,

Long may he reign

May he defend our laws,

And ever give us cause,

To sing with heart and voice,

God save the King

This story originally appeared in Los Angeles Times.
https://724pharmonline.at/index